Fix bug with syncing abilities to a user that has a role (#73)
The SyncAbilities class was attempting to delete role permssions if they were not provided in the new_abilities array. This is not the intended behavior. This class should only modify the abilities on the permissable that is passed in - in this case the user. The user’s role abilities should be left alone.

This pr adds a test to make sure that is the case and modifies the delete_old_abilites method. Previously it was working on the permissables ability array - but that array also can include role abilities if the permissable is a user. Instead switch the class to work on .permissions which are not shared.

This implementation actually has less queries as well.